+++ app/controllers/steps_controller.rb
@@ -34,6 +34,14 @@ class StepsController < ApplicationController
     render @step.contentful_type, locals: {layout: "steps/new_form_wrapper"}
   end
 
+  def edit
+    @journey = Journey.find(journey_id)
+    @step = Step.find(params[:id])
+    @answer = @step.answer
+
+    render "steps/#{@step.contentful_type}", locals: {layout: "steps/edit_form_wrapper"}
+  end
+
   private
 
   def journey_id